# app/models/posts_model.go
package models

// Posts struct to describe posts object.
type Posts struct {
	ID int `db:"id" json:"id" `
	//Title     string    `db:"title" json:"title" validate:"required,lte=255"`
	//Slug      string    `db:"slug" json:"slug" validate:"required,lte=255"`
	//Excerpt   string    `db:"excerpt" json:"excerpt"`
	//Content   string    `db:"content" json:"content"`
	//Type      string    `db:"type" json:"type"`
	//Status    string    `db:"status" json:"status" validate:"required"`
	//Author    int       `db:"author" json:"author" validate:"required"`
	//CreatedAt time.Time `db:"created_at" json:"created_at"`
	//UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
}